Address Kevin's code review comment (nit)
- https://github.com/near/mpc/pull/1813#discussion_r2721919734
